fix(#52): Fix numericality validation bug when using registered attributes and defaults
This change fixes that bug described in #52 by simply not setting a default
in `attributes` and instead, relying on the default behavior from `store`
instead so that we don't have two different mechanisms fighting to set
the default value.

This patch seems to resolve the behavior and correctly returns
`false` to `valid?` in the cases where it should, while still retaining
what we expect the behavior of a `default` to be.

Fixes #52